refactor(ensnode-sdk): change domain-types.ts to types.ts
