+++ packages/ensnode-sdk/src/ens/names.ts
@@ -1,4 +1,4 @@
-import type { Name } from "./domain-types";
+import type { Name } from "./types";
 
 /**
  * Constructs a name hierarchy from a given Name.
