+++ packages/ensnode-sdk/src/ensindexer/config/index.ts
@@ -1,5 +1,5 @@
 export * from "./deserialize";
-export * from "./domain-types";
+export * from "./types";
 export { isSubgraphCompatible } from "./helpers";
 export * from "./serialize";
 export * from "./serialized-types";
