+++ packages/ensnode-sdk/src/ensindexer/indexing-status/index.ts
@@ -1,4 +1,4 @@
 export * from "./deserialize";
-export * from "./domain-types";
+export * from "./types";
 export * from "./serialize";
 export * from "./serialized-types";
