+++ packages/ensnode-sdk/src/shared/index.ts
@@ -1,6 +1,6 @@
 export * from "./cache";
 export * from "./collections";
-export * from "./domain-types";
+export * from "./types";
 export * from "./serialized-types";
 export {
   deserializeBlockRef,
