+++ packages/ensnode-sdk/src/shared/serialize.ts
@@ -1,10 +1,10 @@
-import type { BlockRef, ChainId, Datetime } from "./domain-types";
 import type {
   ChainIdString,
   DatetimeIso8601,
   SerializedBlockRef,
   UrlString,
 } from "./serialized-types";
+import type { BlockRef, ChainId, Datetime } from "./types";
 
 /**
  * Serializes a {@link ChainId} value into its string representation.
